@@ -18,6 +18,7 @@
 		"build": "npm run build:esm && npm run build:cjs",
 		"build:watch": "tsc -w",
 		"clean": "rimraf dist esm",
+		"lint": "eslint .",
 		"test": "jest",
 		"test:watch": "jest --watch",
 		"preversion": "build",
@@ -28,12 +29,15 @@
 	"devDependencies": {
 		"@babel/preset-env": "^7.19.4",
 		"@babel/preset-typescript": "^7.18.6",
+		"@eslint/js": "^9.32.0",
 		"@types/jest": "^29.5.13",
 		"@types/random-seed": "^0.3.3",
+		"eslint": "^9.32.0",
 		"jest": "^29.7.0",
 		"rimraf": "^6.0.1",
 		"ts-jest": "^29.2.5",
-		"typescript": "^5.6.3"
+		"typescript": "^5.8.3",
+		"typescript-eslint": "^8.38.0"
 	},
 	"dependencies": {
 		"random-seed": "^0.3.0"
